+++ Reddit/Easy/making_links/src/lib.rs
@@ -1,20 +1,16 @@
 
-pub fn convert(input: &str) -> String {
-    let letters: Vec<char> = "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz"
+pub fn convert(input: usize) -> String {
+    let symbols: Vec<char> = "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz"
         .chars().collect();
+    let mut input = input;
 
-    let mut input = input.parse::<usize>().unwrap();
-
-    let mut result = String::new();
-
+    let mut result: Vec<usize> = vec![];
     while input > 0 {
-        // println!("{} -> {}", input%62, letters[input%62]);
-        result.push(letters[input%62]);
+        result.push(input%62);
         input /= 62;
     }
 
-
-    result.chars().rev().collect()
+    result.into_iter().map(|c| symbols[c]).rev().collect()
 }
 
 #[cfg(test)]