+++ Reddit/Easy/making_links/src/main.rs
@@ -6,5 +6,5 @@ fn main() {
     let stdin = io::stdin();
     let line: String = stdin.lock().lines().map(|l| l.unwrap()).nth(0).unwrap();
 
-    println!("{}", making_links::convert(&line));
+    println!("{}", making_links::convert(line.parse::<usize>().unwrap()));
 }